+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Renamed AppState to GameState for better clarity
+- Improved readability and versatility of assets with bevy_asset_loader
 
 ## [0.0.12] - 2023-10-07